try to patch delayed twitch mystery sub gifts
The random small delay (up to 0.1 seconds) is to handle duplicate events coming in at the same moment from StreamLabs. In the past this would cause double firing of events as they would both be processed in parallel. This isn't perfect still as the duplicate events could end up with the same delay and thus be processed in parallel, however this is very unlikely.
It does have an issue in that if the Twitch MysteryGiftSub event is delayed after the subscription events then they will be double counted. Set as very small delay to try and avoid this issue. The subscription gift events do come in after the mystery gift event, but it is only a delay of a part of a second. The ideal fix is to move to a event queue processing system, but that's far more work for a dead project.